refactor: vnc_to_fb utils
